state: Separate active-session checks from session updates
Commands, the command-line entry point, and the terminal interface import
the full session module to ask whether a session is active.

Those callers only need to check for the active-session marker file. The
session module also creates and removes sessions, stores snapshots, and
changes repository state, so importing it gives a read-only question a much
larger dependency.

This commit begins separating read-only session queries from state changes
by moving the marker path and active-session check into a dedicated data
module and updating every read-only caller.

The next commit will update session tests and import rules to enforce that
split.
