+++ src/git_stage_batch/commands/block_file.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 import sys
 from contextlib import nullcontext
 
-from ..data.session import path_is_intent_to_add, session_is_active
+from ..data.session import path_is_intent_to_add
+from ..data.session_marker import session_is_active
 from ..data.undo_checkpoints import undo_checkpoint
 from ..data.ignore_files import (
     add_file_to_local_exclude,
