+++ src/git_stage_batch/data/session.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import json
 import os
-from pathlib import Path
 import shutil
 
 from .batch_refs import snapshot_batch_refs
@@ -74,21 +73,6 @@ def _journal_index_entries(
         }
         for file_path in file_paths
     ]
-
-
-def active_session_marker_path(git_dir: Path | None = None) -> Path:
-    """Return the active-session marker path without creating state directories."""
-    state_dir = (
-        git_dir / "git-stage-batch"
-        if git_dir is not None
-        else get_state_directory_path()
-    )
-    return state_dir / "session" / "abort" / "head.txt"
-
-
-def session_is_active(git_dir: Path | None = None) -> bool:
-    """Return whether a batch staging session marker exists."""
-    return active_session_marker_path(git_dir).exists()
 
 
 def _diff_index_name_status(*, intent_to_add_visible: bool) -> dict[str, str]:
